Break Up TextLayoutManager
